Moved EmojiReactionHandler to Quotebot.Interactions.EmoteReaction namespace. Renamed type Quotebot.Interactions.EmoteReaction to QuotedEmoteReactionHandler
